mixnet: Test with multiple auths
Using the mock_query we can now test the mixnet chain call. I've to add
a new field in the Mixnet model because if we use the same server as two
auths we need to differ the position in the chipher chain.

So now it's possible to use the same server as multiple authorities,
although this doesn't give us more security because we do multiple
encryption and decription, but private keys are in the same db so it's
the same that using one auth.