Fix postproc in voting tally
I've fixed the postproc call in the voting tally and also updated the
testvoting django command.

I've also copied the testvoting django command to a django test, now
we've the mods mocked so we can test this with normal tests and this
command is not needed anymore, maybe we should remove in the future
because the test is doing basically the same with this patch.