gateway app to redirect queries
We've the base.mods module to redirect all queries to submodules using
the configuration APIs, but for the voting booth we don't use that, we
were using the final endpoint full url.

To avoid CORS and url problems, this patch adds a new django app that
provides a gateway url to proxy all requests to the corresponding
module, so the request will go to the gateway in the same server and
then this gateway will make the real request to the final servicer and
return the response.

This simplifies the interface to vote and will allow us to create
external tools that doesn't need to know the full server infrastructure,
only the gateway location.