Use jsonnet file for server configuration
The jsonnet is a templating language that let us define a json output
with a simple language. It adds variables, conditionals and other useful
tools to extend the json format.

https://jsonnet.org/

We can use this for the server configuration so we don't need to write
python code and we use the jsonnet format instead.